Fix route_matcher specs in Rails 4
The rails assertion that underlies the `route_matcher` matcher has
changed in Rails 4 to depend on the `message` method from MiniTest.
RSpec is smart enough to make that method available, but only for
controller and routing specs (and only in 2.13.1+).

I marked the `route_matcher` specs as controller specs and updated the
rspec-rails gem development dependency to 2.13.1 or later (but less than 3).